refactored: incident timeline, card & comps
