+++ src/constants/config.js
@@ -1,5 +1,5 @@
 const Config = {
-    API_BASE_URL: 'http://localhost:8000',
-    WEBSOCKET_URL: 'ws://localhost:8000/ws',
+  API_BASE_URL: 'http://localhost:8000',
+  WEBSOCKET_URL: 'ws://localhost:8000/ws',
 };
 export default Config;
