+++ src/lib/websocket.js
@@ -4,32 +4,32 @@ let socket;
 let messageCallback = null;
 
 export function connectWebSocket(orgSlug, onMessageCallback) {
-    socket = new WebSocket(Config.WEBSOCKET_URL +`?org=${orgSlug}`);
+  socket = new WebSocket(Config.WEBSOCKET_URL + `?org=${orgSlug}`);
 
-    messageCallback = onMessageCallback;
+  messageCallback = onMessageCallback;
 
-    socket.onopen = () => {
-        console.log("WebSocket connected");
-    };
+  socket.onopen = () => {
+    console.log('WebSocket connected');
+  };
 
-    socket.onmessage = (event) => {
-        const data = JSON.parse(event.data);
-        if (messageCallback) {
-            messageCallback(data);
-        }
-    };
+  socket.onmessage = (event) => {
+    const data = JSON.parse(event.data);
+    if (messageCallback) {
+      messageCallback(data);
+    }
+  };
 
-    socket.onclose = () => {
-        console.warn("WebSocket disconnected");
-    };
+  socket.onclose = () => {
+    console.warn('WebSocket disconnected');
+  };
 
-    socket.onerror = (err) => {
-        console.error("WebSocket error:", err);
-    };
+  socket.onerror = (err) => {
+    console.error('WebSocket error:', err);
+  };
 }
 
 export function sendMessage(message) {
-    if (socket && socket.readyState === WebSocket.OPEN) {
-        socket.send(JSON.stringify(message));
-    }
+  if (socket && socket.readyState === WebSocket.OPEN) {
+    socket.send(JSON.stringify(message));
+  }
 }
